Migrate core library from graphity to librl.
Since I am using RL for multiple projects, I want to consolidate my code.
`librl` will contain the majority of my 'magic' RL code.

After this commit, I should be able to move the graphity code to librl while preserving history.
