+++ graphity/agent/grad.py
@@ -1,21 +1,17 @@
 import torch
 import torch.nn as nn
-import numpy as np
-from numpy.random import Generator, PCG64, random
 
 # Needed for add_agent_attr() decorator
-import graphity.agent
+import librl.agent
+# Needed to compute grads
 import graphity.grad
 
 # The random agent random selects one edge pair to toggle per timestep.
-@graphity.agent.add_agent_attr()
+@librl.agent.add_agent_attr()
 class GradientFollowingAgent(nn.Module):
     def __init__(self, H, hypers):
         # Must initialize torch.nn.Module
         super(GradientFollowingAgent, self).__init__()
-        # I like the PCG RNG, and since we aren't trying to "learn"
-        # anything for this agent, numpy's RNGs are fine
-        self.rng = Generator(PCG64())
         self.H = H
 
     def act(self, adj, toggles=1):
