+++ graphity/task/__init__.py
@@ -1,3 +1,2 @@
-from .distribution import *
 from .sampler import *
-from .prototypes import*
+from .task import *
