+++ librl/__init__.py
@@ -1,0 +1,2 @@
+def hello_world():
+	print("Hello world")
