+++ requirements.txt
@@ -7,3 +7,4 @@ Pillow==8.0.0
 scipy==1.5.2
 torch==1.6.0
 torchvision==0.7.0
+more-itertools==8.6.0
