+++ setup.py
@@ -6,11 +6,12 @@ setup(name='graphity',
       author='Matthew McRaven',
       author_email='mkm302@georgetown.edu',
       url='https://github.com/Matthew-McRaven/graphity',
-      packages=['graphity',
-      'graphity.agent',
+      packages=['graphity', 'librl',
+      'graphity.agent', 'librl.agent',
       'graphity.environment', 
       'graphity.graph', 'graphity.grad',
-      'graphity.nn',
-      'graphity.replay',  
-      'graphity.task','graphity.train'],
+      'librl.nn',
+      'librl.replay',  
+      'graphity.task', 'librl.task',
+      'graphity.train']
      )
