+++ tools/librl_test.py
@@ -1,0 +1,3 @@
+import librl
+
+librl.hello_world()
